move users to pkg
