+++ admin/auth.go
@@ -7,7 +7,7 @@ import (
 	"github.com/crewjam/saml/samlsp"
 	"github.com/jmpsec/osctrl/admin/sessions"
 	"github.com/jmpsec/osctrl/pkg/settings"
-	"github.com/jmpsec/osctrl/users"
+	"github.com/jmpsec/osctrl/pkg/users"
 	"github.com/rs/zerolog/log"
 )
