+++ admin/handlers/json-carves.go
@@ -7,7 +7,7 @@ import (
 	"github.com/jmpsec/osctrl/carves"
 	"github.com/jmpsec/osctrl/pkg/queries"
 	"github.com/jmpsec/osctrl/pkg/settings"
-	"github.com/jmpsec/osctrl/users"
+	"github.com/jmpsec/osctrl/pkg/users"
 	"github.com/jmpsec/osctrl/utils"
 	"github.com/rs/zerolog/log"
 )
