+++ cli/utils.go
@@ -6,7 +6,7 @@ import (
 	"unicode/utf8"
 
 	"github.com/jmpsec/osctrl/pkg/nodes"
-	"github.com/jmpsec/osctrl/users"
+	"github.com/jmpsec/osctrl/pkg/users"
 	"github.com/jmpsec/osctrl/utils"
 )
